This commit fixes a top-level element bug
If the layout for `<my-component>` has `<my-component>` as its root
element, this is a special case. The semantics of this case are intended
to be the same as a top-level `<div>`.

However, in an attempt to share code, we inadvertantly introduced a bug
where we installed the dynamic attributes defined by JS APIs twice (via
`normalizeComponentAttributes` and post-processing of the parent
template).

The fix, therefore, is to not do anything special with dynamic
attributes and rely on the same post-processing step as the `<div>`
case.