Remove post-processing ("attributes") hook
This was previously used for attaching outer attributes as well as
system attributes (e.g. class="ember-view" id="ember123") onto regular
HTML elements at the root, once they have been rendered.

This results in multiple AttrNodes for a single attribute, which means
that there is no guarantee that any particular AttrNode will actually
"win". This incorrect strategy "works" a surprising amount of the time,
but the failure modes are unacceptable. (See new tests introduced in the
previous commit.)

This is in anticipation of making all top-level elements in a
component's layout dynamic, so that they can share attribute merging
logic with the "identity element" case (top-level `<my-component>` in
the layout for `my-component`).