[BUGFIX release] Remove wasted DK's for `template` and `layout`.
We used to observe and rerender properly when `template` or `layout`
changed. As of 1.13 we do not rerender any longer, and the additional
dependent keys cause extra unusable observers (which add up perf wise).

It is unclear if the change to make `layout` not rerender is a
regression, but we certainly shouldn't be paying for observation when we
can't take advantage of it.