Blocks are no longer shared after createChildScope
Previously, `createChildScope` inadvertantly did not `Object.create` the
parent `scope.blocks`. Originaly, `scope.block` was not an object, but
when `scope.blocks` was introduced, this detail was missed.

The consequence is that it was possible to cause a block invocation to
mutate the `blocks` of a parent scope, changing the behavior of
`{{yield}}` in the parent scope.

We observed an infinite recursion in Ember as a result.