AST transformations can mark a component as static
Today, component nodes are always assumed to be dynamic, meaning that
the starting and ending nodes of their output can change. To understand
why, consider Ember components that emit fragments (“tagless
components”). The nodes represented by the fragment can change,
resulting in a changed start and end node.

This matters because MorphLists assume that their mounted morphs have
stable start and end nodes. If we encounter a dynamic node at the
beginning or end of a template, we insert a static boundary node
(an empty text node) to guarantee static boundaries.

Generally speaking, this strategy is semantics-preserving, because
empty text nodes do not affect the rendering. However, when making
changes to existing constructs, it is important to avoid changing the
exact nodes that get produced (someone may be doing
`component.element.firstChild` for example).

This commit allows an AST plugin to say that a component’s output
will definitely maintain its first and last node after the initial
render. This allows us to remove the boundary nodes reliably in that
scenario.