Templates indicate whether they are empty
The `manualElement` function is designed to take an invocation of a
component (`<my-button></my-button>`) and produce an element that looks
exactly the same as if the user had used a regular element
(`<div></div>`).

One minor difference was that the implementation of `manualElement`
unconditionally inserted a `{{yield}}` inside of the template it builds,
just in case there is some content inside of it. If it turns out that
there is no content inside, the `{{yield}}` still emits an empty
comment, because it has no way to know that the dynamic area it
represents will never have content.

This change is a targeted solution to this problem. It adds an `isEmpty`
flag to templates, and allows code using `manualElement` to skip the
`{{yield}}`. At the moment, this allows libraries to check `isEmpty`
and skip the yield if desired.