Don’t compile templates unconditionally in tests
When running a targeted test, it is useful for the compiler to only run
for the test you are working on.

The previous code unconditionally compiled a lot of templates at the top
level, so even when running a focused test, any debuggers placed in the
compiler would run many times for these top-level tests.

This change makes all compilation in tests happen inside the test
itself, fixing this problem.