Skip AST transform without Glimmer Components
To reduce the possibility of regressions when the
`ember-htmlbars-component-generation` flag is off, this commit moves an
important AST transformation behind the flag.

This AST transformation makes top-level elements in *all* templates
dynamic, and requires runtime work to restore the original semantics.

Moving the AST transformation behind the flag limits the impact of any
bugs in the runtime work.