Disallow fragments in GlimmerComponent for now
There are multiple competing proposals for how to represent fragments in
GlimmerComponents, and we have not yet settled on the exact solution.

In the interim, GlimmerComponents require a top-level wrapper element.
This commit also provides a good error message describing exactly what
the user did wrong. In the future, when fragments are allowed, we can
update the error message to suggest opting into a fragment.