[BUGFIX release] Ensure initializers can augment customEvents.
The initial refactor for Ember 1.11 to use `ApplicationInstance` broke
the ability for an initializer to specify `customEvents`.  This meant
for example that addons could not add their own events to the listing
before the `EventDispatcher` was setup.

The reason for the failure, was that we were copying `customEvents` from
the application upon `init` of the `ApplicationInstance` (which is done
before initializers are called), and never looking at
`application.customEvents` again.

The fix is to avoid eagerly copying from `application.customEvents` when
the application instance is created, and to do it just before setting up
the event dispatcher.

Also, apparently `packages/tests/application_lifecycle.js` test file was
never running because it didn't have the proper test file suffix.