Allow instanceInitializers to set `customEvents`.
Instance initializers were previously not able to affect the
`customEvents` that will be provided to the `EventDispatcher`. This is
most helpful when you are receiving configuration that is registered as
in an initiailzer.

The fix here is to run instance initializers before the event dispatcher
is setup, so that they can augment the `customEvents` hash.