[BUGFIX canary] Honor `customEvents` opting out of event listeners.
Setting `customEvents` to a hash in an application is a way to add new
listeners that the `EventDispatcher` knows about.

Often, users see this and decide that they want to disable a given event
(for example `mouseenter` and `mouseleave` events) and think they can do
something like:

```javascript
var App = Ember.Application.create({
  customEvents: {
    mouseenter: null,
    mouseleave: null
  }
});
```

Unfortunately, that just makes the `EventDispatcher` call a method named
`null` on the view/component instances (LOL).

This fixes that bug by allowing a `null` value to disable the event
listener.