Test all styles of top-level element in <GC>
Previously, we had hardcoded tests for two styles of top-level element
in a GlimmerComponent. For the layout of `<foo-bar>`, we had tests for
the top-level element being `<foo-bar>` itself and being a `<div>`.

A recent commit allowed the top-level element to also be any dasherized
element that was not registered as an Ember component (aka “web
components”). This commit modifies the tests to enumerate all three
styles and runs the same set of tests for each.