Clean up server-side rendering tests
The tests for running Ember applications inside a Node.js environment
evolved organically and were in need of a cleanup. In particular, there
was so much boilerplate code that was repeated for each test that it
was hard for the layperson to follow along, and even experienced users
had trouble understanding what the intent of the test was without
careful parsing.

This commit splits the test into two files:

1. Simple component rendering tests that just verify that a single
   component can be rendered to a string of HTML
2. Integration tests that boot and test an entire application.

The integration tests have been cleaned up significantly. I implemented
a small DSL for building up an application in the tests, as well as some
helper methods that abstract the steps necessary to put Ember into
server-side rendering mode, as well as the asynchrony that comes with
it.

That abstraction happens in a helper that wraps QUnit.module, giving any
tests in that module a pre-built environment for registering classes
in the application, then routing to a URL and testing the resulting
HTML.