[BUGFIX] Disable component lifecycle hooks in FastBoot
This commit fixes #11325, which reported a regression where components
and views lifecycle hooks were inappropriately invoked in FastBoot.

FastBoot needs a mechanism to signal to the rendering layer that it is
operating in an environment without a full DOM, and thus userland hooks
that may expect a fully-functioning DOM (didInsertElement, didRender,
etc) should not be invoked.

Previously, the Renderer’s constructor accepted a second argument that
disabled this, but support for that appears to have been lost somewhere
along the way.

This commit introduces a test that verifies these hooks are disabled.
It also causes the RenderEnv to copy the Renderer’s `destinedForDOM`
flag, making it available inside HTMLBars helpers and keywords.

Now, the ViewNodeManager and ComponentNodeManager check the env for the
`destinedForDOM` flag, and if it’s false, do not invoke these lifecycle
hooks.