[BUGFIX release] fix two bugs in closure actions
Closure actions accidentally invoked `addDependency` for all params and
hash every time the value was recomputed.

Effectively, this means that when the params and hash value change, the
stream representing the closure action gets totally superfluous repeated
dependencies. This probably doesn’t cause incorrect behavior, because
the subscriptions are simply responsible for dirtying nodes, but it does
cause a memory leak.

Additionally, this commit adds a label to the closure action stream,
which was previously missing.