[BUGFIX beta] Fix stack overflow in wrap.
Fix stack overflow when adding terminating wrapper to superFunc in wrap.

For some reason, the recursive call to wrap that doesn’t have an obvious termination, yet wasn’t causing a stack overflow until minified.

Something is wrong with our tests that are supposed to run against the minified build.