[BUGFIX beta] Allow deprecated access to registry.
As part of the `ember-registry-container-reform` the registry was moved
from `Application#registry` / `ApplicationInstance#registry` to
`__registry__`, and a new public API was created allowing access to the
majority of functions that folks used directly on `Application` or
`ApplicationInstance` instances.

We created an alias `container` property for `ApplicationInstances`,
but missed creating the same thing for `registry` which was exposed
non-underscored on both `Application` and `ApplicationInstance`
instances.

This adds an alias for all public `registry` methods at their original
location and provides a helpful deprecation warning instructing you what
to use instead.

Also, changes the `ApplicationInstance` `Object.defineProperty` calls to be
done once on the `ApplicationInstance` prototype (instead of on each instance).

(cherry picked from commit 1d40a0b98c1e10ee56ef7499d01053dd9ca3da3e)