[BUGFIX release] Ensure `_actions` specified to extend works.
The initial deprecation of `_actions` and moving it to `actions` was a
bit too naive. The `deprecateProperty` helper was used, which meant that
when `_actions` was set, it would call `set(this, 'actions', newValue)`.
This meant that passing `_actions` to `Ember.Route.extend` (or
`.reopen`) resulted in the `_actions` completely clobbering `actions`
(and loosing any previously defined actions there).

The fix was to:

* Replace the generic `deprecateProperty` usage with a custom
 `Object.defineProperty` call, that defines a setter with an assertion
 instead of clobbering.
* Implementing a `willMergeMixin` that mutates the `props` at extend
  time (to ensure that `_actions` is moved to `actions` with a
  deprecation).

(cherry picked from commit dcc0dc3d8a3b9546c6da4cc8b2cb6086a3c98a00)