[BUGFIX beta] Avoid setting shadowed properties twice.
The recent work done on the AttrsProxy means that everything in
`this.attrs` is propagated back to the root of the component for every
time `willReceiveAttrs` is called. Unfortunately, this was also being
done in the various node-managers causing the same properties to be set
twice for each rerender.

(cherry picked from commit 8d4f0a594dddf2f0954b6b396a22cb144c90da51)