Remove most of AttrsProxy
Previously, we tried to restrict which attributes we set eagerly, and
clean up the rest using unknownProperty. However, over time, making that
work correctly has gotten more and more complicated, and the cost of the
laziness is more than the benefits.

This commit goes back to an eager setting strategy, eliminates
unknownProperty and the reliance on _internalDidReceiveAttrs. It leaves
a setter interceptor for propagating sets to mutable cells.