[BUGFIX beta] Provide a helpful error for undefined closure actions.
In the following scenario:

```handlebars
{{#foo-bar blah=(action somethingThatIsUndefined)}}
  Click Me!
{{/foo-bar}}
```

Previously, an error was only thrown when attempting to invoke
`this.attrs.blah()`.  After this change a helpful error like the following
will be thrown:

```
An action could not be made for `somethingThatIsUndefined` in
< component object id>.  Please confirm that you are using either
a quoted action name (i.e. `(action 'somethingThatIsUndefined')`)
or a function available in < component object id>.
```