[BUGFIX canary] Trust Ember.set
This change eliminates unnecessary dirtying of proxy streams when the
source object is an Ember.Object. It results in significant re-render
performance improvement in certain scenarios, especially when people are
iterating over lists of Ember.Objects to render components with
expensive lifecycle hooks.

We already have a longstanding convention that people need to use `set`
to mutate Ember.Objects if they expect any observability, and as long as
people are using `set`, interior mutability will be captured by other
observers further downstream.

The same is not necessarily true for POJOs. We would like to allow
something conceptually like:

    myComponent.set('thing', aPojo);
    aPojo.deep.inside.here = 'newValue';
    myComponent.notifyPropertyChange('thing');

to work correctly, such that the deep mutation gets noticed. Which is
why this change treats POJOs and Ember.Object's differently.