[BUGFIX release-1-13] Ensure concat streams unsubscribe properly.
The stream returned by `concat` should subscribe to any input
streams so that the "concat stream" is marked as dirty if any of
the input streams are dirtied. This is/was working properly.

Unfortunately, when `subscribe`ing directly as done in 1.13.0, the
subscriptions are never cleared if the "concat stream" is deactivated.
This means, that listeners build up over time and leak the "concat
streams" and the `KeyStream`'s / `ProxyStream`'s that are in use.

Using `stream.addDependency` instead of manually `subscribe`ing allows
the returned "concat stream" to properly remove subscribers when it is
deactivated.

(cherry picked from commit dc4fbd9e5c7e1b84f2e9fb3d112bfd83f710ff1a)