[BUGFIX beta] Avoid extra work for `Ember.Component#layout`.
---

Prior to this change, `layout` was a CP that did a few things:

* `get(this, 'layoutName')` (and if set `this.container.lookup`)
* `get(this, 'defaultLayout')`

After this change, `layout` is a simple property that avoids that extra
work.

---

Prior to this change, `defaultLayout` would actually override a layout
that came from the container, this was incorrect (literally the opposite
of what we wanted).

This corrects that logic, and ensures that we only attempt to use the
`layout` property if the layout was not found during the container
lookup (which properly mimics the behavior in 1.0 - 1.12 where `layout`
was injected into the component if a layout was found in the container).