Reorganize
Among other things, this makes it a lot clearer when work is happening,
and gives callers more control over whether certain work is actually
needed.

The old render/block process lumped a lot of stuff together (including
scope binding) that really should be the responsibility of the code
setting up the block or calling render.