Ensure attributeBindings work when legacy view addon is disabled.
When the `_Ember.ENV._LEGACY_VIEW_SUPPORT` flag is falsey, the `view`
stream is not setup as of 6963750. This resulted in anything added in
`attributeBindings` being `undefined` (because the attribute binding
system was essentially using `href=view.href` instead of
`href=this.href` or just `href=href`).

You might thing (as I did initially): "HOW CAN WE NOT HAVE A SINGLE TEST
THAT CHECKS FOR AN ATTRIBUTE VALUE?!?!?!?", and the answer is somewhat
more disconcerting.  We currently run all tests with the legacy view and
controller flags set to `true`. :(

Future test refactoring efforts will be needed to fix that mistake.