[BUGFIX beta] Make `{{get something 'path.goes.here'}}` work.
The `get` keyword should be roughly the same as `Ember.get`, but
previously we were asserting when the path included a `.`. This was most
likely because in the non-dynamic key case we were always using a
`KeyStream` instead of calling `source.get(newPath)` (a `KeyStream` will
assert when its path includes a `.` but `BasicStream#get` will
automatically create the correct stream for that path).

This removes the asserts, hard coded `KeyStream` usage, and adds a
number of tests for nested keys.