Refactor application (instance) boot paths
This is a dialed back attempt at making the `ApplicationInstance`
boot process fully async. I wrote all the code to make that happen,
only to find out that a lot of tests implicitly depend on the process
being synchronous. So, I did the following instead:

1. Create an async version of `boot` on both `Application` and on
   `ApplicationInstance`. Going forward, new APIs (`visit`, etc)
   should use the async version.

2. Keep an internal `_bootSync` method on both classes and use them
   in places where we rely on the process being synchronous.

In the short term, having to support both synchronous and
asynchronous might make things look a bit more complicated, but it
will help us iterate towards where we want to be.

Separately, this commit also solidified the concept of a `boot`
phase on `ApplicationInstances`. Instead of overriding things on
the instance in an ad-hoc manner and implicitly rely on the timing
of these calls (e.g. `overrideRouterLocation` must be called before
the router is instantiated/cached), they have now been consolidated
into options that you pass into the `boot()` method.

Having an explicit boot phase makes it clear what can and cannot
happen at runtime.

(TODO: move the "start routing" logic into the boot phase)