Update QUnit test timeout to 15 seconds.
We have an annoying intermitent test failure that occurs when running
tests on older browsers in Sauce Labs. The failure is generally
occurring when the VM itself is slower than normal, and will pass after
a restart.

This increases the timeout to allow more time (from 7 seconds to
15 seconds) in the hopes that this resolves the majority of our random
failure scenarios.