Avoid timing brittleness in Ember.Select test.
The goal of this test is to ensure that the following scenario:

1. `selection` is set to a slow resolving promise
2. `selection` is set to another value (a promise that resolves before
  the "slow resolving" promise above)

Results in the newest assigned selection (2. from above) and NOT the
most recently resolved promise (1. from above).

---

This test randomly fails in Sauce Labs in the previous rendition,
because the VM sometimes runs slowly and the 10ms difference between the
two `run.later`'s is lost.