[BUGFIX canary] Add meta and GUID keys to stream shape.
After `init` is ran for stream objects we seal the object to ensure that
we maintain a consistent object shape.

In some scenarios the Ember Inspector needs to call `Ember.guidFor` on a
stream to keep track of which streams have already been seen/walked.
Unfortunately, we forgot to add `GUID_KEY` and `__ember_meta__` to the
base stream object before sealing, which causes an error:

```javascript
Uncaught TypeError: Cannot define property:__ember1442443003034, object is not extensible.
```

This adds the proper placeholders to the base stream class to prevent
this error.