[BUGFIX canary] Ensure components actions function without controller.
On canary we no longer create `controller` and `view` template locals.
Unfortunately, our test suite still pretends that they are there (since
they are added back in by the legacy addons) so all tests pass that were
using `scope.getLocal('controller')` and friends).

This commit adds a test for sending actions that runs without the flags
enabled (the test failed prior to this change), and updates
component-node-manager to provide either `_controller` *OR*
`_targetObject` to the component that is being created.