[PERF] Bail on KeyStream#compute if bool or string
When compute is called it is expected that object is an actual Object,
sometimes the object is a Bool or String and yet we continue down the stack look
values up on it even though the result will be undefined. This change
bails out early if `object` is a Bool or String.