[BUGFIX release] Avoid `this.attrs.params` in LinkComponent.
Due to the nature of `positionalParams`, it is trivial to supply
`params` as a hash argument to `{{link-to}}` instead of ordered args
(this allows truly dynamic link's). However, if `{{link-to params=foo}}`
is used `params` will be a `MutableCell` and the current code throws an
error in that circumstance.

*tldr;* `this.attrs` should be avoided. It is quite a :troll: in
non-`GlimmerComponent`'s.