[BUGFIX Release] allow watching of ES5+ Getter
Before this commit MandatorySetter enabled + current property being watched
resulted in `get` always fetching via meta.values. Although this typically works,
the default getter for a MandatorySetter already does this, so it is actually redundant.

Going further, when a user defines their own getter this code-path is incorrectly taken,
instead the original value (which originated from the getter) should be the consulted value