[FEATURE ember-htmlbars-local-lookup]
Implements local lookup concepts. This new concept is primarily
used from the template layer in this initial implementation,
but the `source` option is usable generally.

There are a couple main concepts being added here:

* Expanding a `type:name` relative to a given source `type:name` into a
  fully qualified `type:name` (that includes the source info).
* Expose the currently rendering `template`'s `meta` information on the
  `RenderEnv`.
* Adding a second layer of lookups in all places that components or
  helpers are resolved.

\#### `expandLocalLookup`

We use the `expandLocalLookup` function on the resolver (when present)
to take both the source and target full names and return a new fully
qualified full name. By default the resolver will not have an
`expandLocalLookup` function, but an example implementation exists in
the `ember-htmlbars/tests/integration/local-lookup-test` module here.

\#### `meta`

As of Ember 1.12, metadata has been appended to every template to allow
better customization of things like deprecation messages and template
compilation errors.

With the changes in this commit that metadata is exposed on the
`RenderEnv` which is present through the various internal glimmer engine
hooks (and provides us the `source` information needed for the local
lookups added here).

\#### local + global lookups

The various places that are responsible for looking helpers and
components up have been updated to do two layers of lookups.  Once for
the current global style lookups, and a second for lookups relative to
the invocation template.