[FEATURE ember-container-inject-owner] Inject fake container for non-extendable factories.
The fake container is required to provide access to the `container`
during instantiation of objects by non-extendable factories.

The fake container provides access to the following methods:

* `lookup`
* `lookupFactory`

Deprecation warnings are provided that direct the user to call
`getOwner` followed by the equivalent `ContainerProxy` methods
(`lookup` and `_lookupFactory`) on the owner.