[BUGFIX beta] Ensure Mixin.prototype.toString is defined.
`Mixin.prototype` is sealed after its initial definition, but
`Mixin.prototype.toString` is assigned later in `ember-runtime/namespace`.
This causes issues when `ember-template-compiler.js` is loaded before
the rest of Ember (because the `Mixin.prototype` is sealed).
Unfortunately, the debug functions (for `Object.seal`) does not work in
the `ember-metal` package at this time.

This fix is a temporary fix to allow the loading of
`ember-template-compiler.js` to be before or after `ember.debug.js`. The
more permanent fix is to ensure that `debugSeal` works in ember-metal
properly.

(cherry picked from commit 895bbd7f4e496353d4f1aac331ff7bbd15f8937a)